refactor(dashboard-pages): extract DashboardShell, SettingsShell, settings groups
Moves the dashboard and settings layout shells into `@getmunin/dashboard-pages`
so downstream consumers (munin-cloud) can compose the same structure instead
of redeclaring it. Adds `OSS_SETTINGS_GROUPS` and `extendSettingsGroups()` so
extra nav items can be merged into the OSS defaults. The web app's
`dashboard/layout.tsx` and `dashboard/settings/{layout,page}.tsx` collapse to
thin call sites; the standalone `nav-config.ts` is deleted.
