+++ apps/web/app/[locale]/dashboard/settings/page.tsx
@@ -1,11 +1,4 @@
-import { redirect } from '@/i18n/navigation';
+import { createSettingsIndexRedirect } from '@getmunin/dashboard-pages';
 import { DEFAULT_LOCALE } from '@/i18n/locales';
 
-export default async function SettingsIndexPage({
-  params,
-}: {
-  params: Promise<{ locale: string }>;
-}) {
-  const { locale } = await params;
-  redirect({ href: '/dashboard/settings/team', locale: locale || DEFAULT_LOCALE });
-}
+export default createSettingsIndexRedirect({ defaultLocale: DEFAULT_LOCALE });
